# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Стильный призыв API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/me:
    get:
      operationId: getMyProfile
      tags: [user-profiles]
      summary: Get current user's profile (creates default on first access)
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [user-profiles]
      summary: Update current user's nickname and/or avatarUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "429":
          description: Nickname cooldown not expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
  /profile/{clerkUserId}:
    get:
      operationId: getUserProfile
      tags: [user-profiles]
      summary: Get another user's public profile
      parameters:
        - name: clerkUserId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"

  # ── Admin avatar ───────────────────────────────────────────────────────
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlInput"
      responses:
        "200":
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlPayload"
components:
  schemas:
...
    ErrorPayload:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserProfile:
      type: object
      required: [id, clerkUserId, nickname, nicknameUpdatedAt, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        nickname: { type: string }
        avatarUrl: { type: ["string", "null"] }
        nicknameUpdatedAt: { type: string }
        createdAt: { type: string }
    UserProfileUpdate:
      type: object
      properties:
        nickname: { type: string, minLength: 1 }
        avatarUrl: { type: ["string", "null"] }

    # ── Messages ───────────────────────────────────────────────────────
...
    RequestUploadUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name: { type: string }
        size: { type: integer }
        contentType: { type: string }
    RequestUploadUrlPayload:
      type: object
      required: [uploadURL, objectPath, metadata]
      properties:
        uploadURL: { type: string }
        objectPath: { type: string }
        metadata:
          type: object
          required: [name, size, contentType]
          properties:
            name: { type: string }
            size: { type: integer }
            contentType: { type: string }