# lib/api-zod/src/generated/types/requestUploadUrlInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Стильный призыв API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestUploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}